automatic batch size for `dp test` (#1165)
* automatic batch size for `dp test`

Resolves #1149.

We start nbatch * natoms from 1024 (or we can set a different number), and iteratively multiply it by 2 until catching the OOM error.

A small issue is that it's a bit slow to catch the TF OOM error. It's a problem of TF and I don't know how to resolve it. Luckily we only need to catch once.

* replace `execuate` with `execute`

* add unittest; bugfix