Fix dplr error (#2436)
DPLR is broken due to the change of coeff(int narg, char **arg)
function. The current coeff function requires that the number of types
in LAMMPS no more than that in model. In dplr, however, there might be
more types in LAMMPS than in the model. This check should be bypassed.

---------

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>
Co-authored-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>