migrate Python build backend to scikit-build-core (#2798)
This PR migrates the Python build backend from scikit-build to
[scikit-build-core](https://github.com/scikit-build/scikit-build-core),
which is being actively developed.

There are no breaking changes to users; all tests should be passed
without modification.

The behavior for editable installation (`pip install -e .`) is changed.
Before, the compiled binary libraries (e.g. `libdeepmd_op.sp`) are
installed into the source directory. In this PR, the library files are
installed into `site-packages` separately, while the Python files are
still in the source code and can be modified for development.

---------

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>