docs: document horovod on Conda-Forge (#3001)
Recently I contributed Horovod to conda-forge at
https://github.com/conda-forge/staged-recipes/pull/24472. I have tested
it with DeePMD-kit. Now the Conda-forge channel has all the features in
our deepmodeling channel and even more available platforms.

I wonder if we should recommend Conda-forge in favor of our DeepModeling
channel and use it for the offline installer. In 2019, Conda-forge
lacked support for TensorFlow, so we hosted and built our own channel
against the Anaconda channel. But now Conda-forge has become a mature
channel and has better support for TensorFlow than the Anaconda channel.

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>