fix: invalid read and write when natom grows (#3031)
This PR fixes #3015, which reported an invalid memory read and write
when running `fix deposition` in Lammps to add atoms into the system.
The issue could not be reproduced when `atomic` was removed from the
`pair_style deepmd` line.

As suggested by @njzjz , the possible cause of the bug was the
allocation of `stdfsend`, `stdfrecv`, etc., which only happened when
`out_each == 1`, set by `atomic` in the `pair_style deepmd` command.
Moreover, their size was only determined in the initialization step,
based on the initial `natoms`. The valgrind log also supported this
hypothesis, as it showed `MPI_Gather` and `MPI_Gatherv` errors in the
traceback.

Therefore, we decided to destroy and recreate the memory when
`atom->natoms` became larger than `stdfsend`. This should prevent the
invalid memory access and resolve the issue.