cc: refactor DeepPotModelDevi, making it framework-independent (#3134)
Refactor `DeepPotModelDevi` as a step of #3122. Now, it is just a
wrapper of multiple `DeepPot` classes. Models can have different
behaviors inside different `DeepPot`.

One may argue that the new class needs to prepare the input multiple
times. However, it's not expensive only to copy the memory. Also, during
the simulations, usually we run it every 100 steps.