pt: avoid torch.tensor(constant) during forward (#3421)
`torch.tensor(constant)` copies memory from the CPU to the GPU, so it is
host blocking and should be avoided in the `forward` method.


Before, the CPU waited for the GPU using `cudaStreamSynchronize`,
blocking the CPU from doing the following things, where the CPU memory
needs to be copied to the GPU, a.k.a. host-to-device (H2D).


![1709693858444](https://github.com/deepmodeling/deepmd-kit/assets/9496702/e6fb6281-245f-4620-82bd-dbcd02121e32)

After this PR, all ops in the energy loss are asynchronous, as no H2D
happens.

![1709694622120](https://github.com/deepmodeling/deepmd-kit/assets/9496702/172e1601-1a9c-4236-a1e2-a749edc25c50)

---------

Signed-off-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>